# web/src/lib/schemas/openapi-v3.yml
openapi: 3.1.0
info:
  title: Talaria V3 entity projections
  version: 1.0.0
paths:
...
  /api/v1/claims/{claim_id}/intuition:
    get:
      summary: Intuition readiness for a soft claim
      description: |
        Read-only. ready means a published triple termId exists.
        counter_term_id is resolved on the client via MultiVault getInverseTripleId.
        Does not create triples.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, enum: [ready, not_published]}
                  claim_id: {type: string, format: uuid}
                  chain_id: {type: integer}
                  triple_term_id: {type: string}
                  counter_term_id: {type: [string, 'null']}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/claims/{claim_id}/signals/sync:
    post:
      summary: Verify a user-signed MultiVault deposit and persist Talaria stance
      description: |
        Session required. Backend verifies receipt, MultiVault, receiver wallet, and termId.
        Idempotent on tx_hash. Uncertain is not accepted here.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, tx_hash]
              properties:
                action: {type: string, enum: [support, dispute]}
                tx_hash: {type: string}
      responses:
        '200':
          description: Existing tx already synced
        '201':
          description: Stance created after verified deposit
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
        '409': {$ref: '#/components/responses/StructuredError'}
  /api/v1/claims/{claim_id}/sources:
    get:
      summary: List Scholar corpus documents attached to an Agora claim
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/ClaimSource'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
    post:
      summary: Attach an indexed corpus document to a claim
      description: Session required. Idempotent. No free-text URLs.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [corpus_document_id]
              properties:
                corpus_document_id: {type: string, format: uuid}
      responses:
        '200': {description: Source already attached}
        '201': {description: Source attached}
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/claims/{claim_id}/arguments:
    get:
      summary: List structured arguments targeting a claim
      description: Uses soft_claim_relations (supports, contradicts, qualifies). Off-chain.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 50}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Argument'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
    post:
      summary: Create a sourced argument (or counter-claim) as a soft_claim
      description: Session required. Requires corpus_document_id and fragment_id or quote. No Intuition publish.
      parameters:
        - name: claim_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [relation, statement, corpus_document_id]
              properties:
                relation: {$ref: '#/components/schemas/ArgumentRelation'}
                statement: {type: string, minLength: 1, maxLength: 4000}
                corpus_document_id: {type: string, format: uuid}
                fragment_id: {type: [string, 'null'], format: uuid}
                quote: {type: [string, 'null'], minLength: 1, maxLength: 4000}
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                required: [argument]
                properties:
                  argument: {$ref: '#/components/schemas/Argument'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
  /api/v1/comments/{comment_id}/promote:
    post:
      summary: Promote an owned comment into a structured argument
      description: Comment body is unchanged. User must still supply source and evidence.
      parameters:
        - name: comment_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [relation, corpus_document_id]
              properties:
                relation: {$ref: '#/components/schemas/ArgumentRelation'}
                statement: {type: string}
                corpus_document_id: {type: string, format: uuid}
                fragment_id: {type: [string, 'null'], format: uuid}
                quote: {type: [string, 'null']}
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                required: [argument]
                properties:
                  argument: {$ref: '#/components/schemas/Argument'}
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
...
components:
...
  responses:
...
    StructuredError:
      description: Code + message, never raw SQL
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code:
                    type: string
                    enum:
                      - unauthenticated
                      - invalid_action
                      - invalid_target_type
                      - invalid_action_for_target
                      - target_not_found
                      - interaction_not_found
                      - invalid_visibility
                      - batch_too_large
                      - database_error
                      - invalid_address
                      - conflict
                      - expired_challenge
                      - invalid_signature
                  message: {type: string}
  schemas:
...
    ArgumentRelation:
      type: string
      enum: [supports, contradicts, qualifies]
    ClaimSource:
      type: object
      required: [corpus_document_id, title, source_kind, document_type, created_at]
      properties:
        corpus_document_id: {type: string, format: uuid}
        title: {type: string}
        source_kind: {type: string}
        canonical_url: {type: [string, 'null']}
        document_type: {type: string}
        created_at: {type: string, format: date-time}
    Argument:
      type: object
      required: [id, target_claim_id, relation, statement, origin, contribution_status, created_at, sources, evidence]
      properties:
        id: {type: string, format: uuid}
        target_claim_id: {type: string, format: uuid}
        relation: {$ref: '#/components/schemas/ArgumentRelation'}
        statement: {type: string}
        origin: {type: string, enum: [pipeline, user]}
        contribution_status: {type: string, enum: [proposed, active, rejected]}
        created_at: {type: string, format: date-time}
        author:
          type: [object, 'null']
          properties:
            id: {type: string, format: uuid}
            display_name: {type: [string, 'null']}
        sources:
          type: array
          items: {$ref: '#/components/schemas/ClaimSource'}
        evidence:
          type: array
          items:
            type: object
            properties:
              id: {type: string, format: uuid}
              quote: {type: [string, 'null']}
              locator: {type: [string, 'null']}
              source_system: {type: string}
...